me_conceito_api/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
django-admin-interface>=0.26.0
django-colorfield>=0.4.4
Pillow>=12.0
